Refactor create link modal
